refactor: use Authentication as component name
