+++ packages/api-reference/src/components/Content/Authentication/Authentication.vue
