+++ packages/api-reference/src/components/Content/Authentication/index.ts
@@ -1,0 +1,1 @@
+export { default as Authentication } from './Authentication.vue'
