+++ packages/api-reference/src/components/Content/Content.vue
@@ -8,11 +8,11 @@ import { useTemplateStore } from '../../stores/template'
 import type { Spec, Tag } from '../../types'
 import { FlowIcon } from '../Icon'
 import { SectionContainer } from '../Section'
+import { Authentication } from './Authentication'
 import EndpointsOverview from './EndpointsOverview.vue'
 import Introduction from './Introduction'
 import Models from './Models.vue'
 import ReferenceEndpoint from './ReferenceEndpoint'
-import { SecuritySchemes } from './SecuritySchemes'
 import Spinner from './Spinner.vue'
 
 const props = defineProps<{ ready: boolean; spec: Spec }>()
@@ -65,7 +65,7 @@ const moreThanOneDefaultTag = (tag: Tag) =>
       <Introduction
         :info="spec.info"
         :servers="localServers" />
-      <SecuritySchemes :spec="spec" />
+      <Authentication :spec="spec" />
       <template
         v-for="(tag, index) in spec.tags"
         :key="tag.id">
@@ -161,3 +161,4 @@ const moreThanOneDefaultTag = (tag: Tag) =>
   }
 }
 </style>
+./Authentication
