--- packages/api-reference/src/components/Content/SecuritySchemes/index.ts
@@ -1,1 +1,0 @@
-export { default as SecuritySchemes } from './SecuritySchemes.vue'
