+++ packages/api-reference/src/helpers/hasSecuritySchemes.ts
@@ -1,6 +1,6 @@
 import { type Spec } from 'src/types'
 
-export const hasSecuritySchemes = (spec: Spec) => {
+export const hasSecuritySchemes = (spec?: Spec) => {
   // TODO: Show security schemes
   if (Object.keys(spec?.components?.securitySchemes ?? {}).length) {
     return true
